requirements.txt
mstrio-py==11.3.9.101
urllib3